+++ lobster_common/exceptions.py
